requirements.txt
confluent-kafka==2.3.0
psycopg2-binary==2.9.9
apache-airflow==2.10.4
pyspark==3.5.0
praw==7.8.1